modified for the latest error page test

+++ src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -33,6 +33,9 @@ public class GradebookController {
 
 	@GetMapping("/delete/student/{id}")
 	public String deleteStudent(@PathVariable Integer id, Model m) {
+		if (!studentService.studentExistsById(id)) {
+			return "error";
+		}
 		studentService.deleteStudentById(id);
 		m.addAttribute("students", studentService.findAll());
 		return "index";
